Move funs to main.py and add tests
