+++ sqlrunner/main.py
@@ -1,4 +1,43 @@
 # import pymssql
+import argparse
+import csv
+import pathlib
+
+
+def parse_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--database_connection",
+        required=True,
+        type=str,
+        help="String of database connection",
+    )
+    parser.add_argument(
+        "--sql_query",
+        required=True,
+        type=pathlib.Path,
+        help="Path to the SQL file containing the query",
+    )
+    parser.add_argument(
+        "--output",
+        required=True,
+        type=pathlib.Path,
+        help="Path to the output CSV file",
+    )
+    return parser.parse_args()
+
+
+def read_sql(sql_query):
+    """Open and read SQL file"""
+    return sql_query.read_text(encoding="utf-8")
+
+
+def write_results(results, path):
+    """Write results"""
+    with path.open(mode="w", newline="", encoding="utf-8") as output:
+        writer = csv.DictWriter(output, fieldnames=results[0].keys())
+        writer.writeheader()
+        writer.writerows(results)
 
 
 def run_sql(database_connection, sql_query):
